\name{mgm_init}
\alias{mgm_init}
\title{Initialize causalMGM}
\description{Loads the rJava library, tetradLite.jar, and initializes the JVM}
\usage{
mgm_init()
}
\arguments{}
\value{
Does not return any value
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Neha Abraham
}
\note{
Must be the first function called after loading the causalMGM library
}

\examples{
mgm_init()
}
